requirements.txt
asttokens==2.0.5
azure-core==1.23.0
azure-cosmos==4.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
flash==1.0.3
Flask==2.1.2
flask-paginate==2022.1.8
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.1
jupyter-core==4.10.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
nltk==3.7
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.12.0
pymongo==4.1.1
pyparsing==3.0.8
python-dateutil==2.8.2
pywin32==304
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tornado==6.1
tqdm==4.63.0
traitlets==5.2.0
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.1.2
Whoosh==2.7.4
